Policies successfulyy integrated

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
+use Illuminate\Support\Facades\Auth;
 
 use function PHPUnit\Framework\isNull;
 
@@ -16,7 +17,10 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::all();
+        
+        $this->authorize('viewAny', Post::class);
+
+        $posts = Post::where('user_id', Auth::user()->id)->get();
         return view('resources.post.index', ['posts' => $posts]);
     }
 
@@ -35,6 +39,7 @@ class PostController extends Controller
     {
         
         Post::create([
+            'user_id'=> Auth::user()->id,
             'subject'=> $request->subject,
             'post'=> $request->post,
             'status'=> (is_null($request->status)? 0 : 1)
@@ -47,6 +52,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        $this->authorize('view',$post);
         return view('resources.post.show', ['post' => $post]);
     }
 
@@ -55,6 +61,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('view',$post);
         return view('resources.post.edit', ['post' => $post]);
     }
 
@@ -63,7 +70,9 @@ class PostController extends Controller
      */
     public function update(UpdatePostRequest $request, Post $post)
     {
+        $this->authorize('view',$post);
         $post->update([
+            'user_id'=> Auth::user()->id,
             'subject'=> $request->subject,
             'post'=> $request->post,
             'status'=> (is_null($request->status)? 0 : 1)
@@ -76,7 +85,13 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('view',$post);
         $post->delete();
         return redirect()->route('post.index')->with('message', 'Post Successfully Deleted!!');
+    }
+    
+    public function postIndex(){
+        $posts = Post::where('status', 1)->get();
+        return view('pages.index', ['posts' => $posts]);
     }
 }
